--- app/_components/ProjectList/Card/Card.tsx
@@ -1,28 +1,0 @@
-import Image, { StaticImageData } from "next/image";
-import HoverCard from "./HoverCard";
-
-interface CardProp {
-  id: number;
-  titleImage: StaticImageData | string;
-  stackList: string[];
-  wishCount: number;
-  isWish: boolean;
-  projectName: string;
-  subDescription: string;
-  description: string;
-}
-
-function ProjectCard({ project }: { project: CardProp }) {
-  return (
-    <div className="flex flex-col gap-2.5">
-      <div className="group relative aspect-square min-h-[220px] min-w-[220px] overflow-hidden rounded-md">
-        <div className="absolute inset-0 z-30 opacity-0 transition-opacity duration-200 group-hover:opacity-100">
-          <HoverCard isWishProject={project.isWish} wishCount={project.wishCount} stackList={project.stackList} />
-        </div>
-        <Image fill src={project.titleImage} alt={project.subDescription} />
-      </div>
-    </div>
-  );
-}
-
-export default ProjectCard;
