+++ app/_components/ProjectList/ProjectCard/HoverCard.tsx
@@ -1,0 +1,32 @@
+import WishCount from "../../WishCount/WishCount";
+import StackBadge from "./StackBadge";
+
+function HoverCard({
+  stackList,
+  wishCount,
+  isWishProject,
+}: {
+  stackList: string[];
+  wishCount: number;
+  isWishProject: boolean;
+}) {
+  return (
+    <div className="absolute inset-0 z-30 opacity-0 transition-opacity duration-200 group-hover:opacity-100">
+      <div className="flex p-6">
+        <div className="z-20">
+          <div className="grid grid-cols-3 gap-2.5">
+            {stackList.map((stack, index) => (
+              <StackBadge key={index} stack={stack} />
+            ))}
+          </div>
+          <div className="absolute bottom-5 right-5 text-white">
+            <WishCount isFavorite={isWishProject} wishCount={wishCount} />
+          </div>
+        </div>
+        <div className="absolute inset-0 z-10 bg-gradient-to-t from-black to-[rgba(0,0,0,0)] opacity-90" />
+      </div>
+    </div>
+  );
+}
+
+export default HoverCard;
